Moved logic from the ckeditor-webpack-plugin to resolveImportPath.
